# docs/swagger.yaml
basePath: /
definitions:
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ProjectStats:
    properties:
      total_projects:
        type: integer
    type: object
  models.Task:
    properties:
      assigned_to:
        type: string
      assigned_to_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        description: low, medium, high, urgent
        type: string
      project_id:
        type: string
      status:
        description: todo, in-progress, review, done
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.TaskStats:
    properties:
      statuses:
        additionalProperties:
          type: integer
        type: object
      total_tasks:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Production-grade Project Management API with collaboration features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DevPulse API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
  /api/auth/logout:
    post:
      description: Clear the authentication cookie
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Logout
      tags:
      - auth
  /api/auth/register:
    post:
      parameters:
      - description: Registration info
        in: body
        name: input
        required: true
        schema:
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
  /api/projects:
    get:
      description: List projects for the current user or search by query
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project for the current user
      parameters:
      - description: Project info
        in: body
        name: project
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new project
      tags:
      - projects
  /api/projects/{id}:
    get:
      description: Get a single project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get project details
      tags:
      - projects
  /api/projects/stats:
    get:
      description: Get count of projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectStats'
      summary: Get project stats
      tags:
      - projects
  /api/tasks:
    get:
      description: List all tasks or filter by project, search query, or priority
      parameters:
      - description: Project ID filter
        in: query
        name: project_id
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      - description: Priority filter
        in: query
        name: priority
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task within a project
      parameters:
      - description: Task info
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new task
      tags:
      - tasks
  /api/tasks/{id}:
    delete:
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a single task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get task details
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update task details (title, description, status, priority, assignment)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated task info
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Update task
      tags:
      - tasks
  /api/tasks/export:
    get:
      description: Download all tasks in CSV format
      produces:
      - text/csv
      responses:
        "200":
          description: CSV data
          schema:
            type: string
      summary: Export tasks to CSV
      tags:
      - tasks
  /api/tasks/stats:
    get:
      description: Get various statistics about tasks
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskStats'
      summary: Get task statistics
      tags:
      - tasks
  /api/users:
    get:
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: List all users
      tags:
      - users
  /api/users/profile:
    delete:
      description: Delete the authenticated user account
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Delete current user profile
      tags:
      - users
    get:
      description: Get the profile for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile for the authenticated user
      parameters:
      - description: Profile update info
        in: body
        name: profile
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"